# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee",
        "description": "This is the Employee's API",
        "termsOfService": "https://treblalabs.com/",
        "contact": {
            "email": "mail@treblalabs.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
        "https://treblalabs.com": {},
        "/api/funcionarios": {
            "get": {
                "tags": [
                    "Funcionarios"
                ],
                "description": "Get all users from API",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "items_per_page",
                        "in": "query",
                        "description": "Itens per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Form request validation error. Invalid input."
                    }
                }
            },
            "put": {
                "tags": [
                    "Funcionarios"
                ],
                "description": "Store a new user into database",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "nome",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "imagem",
                        "in": "query",
                        "description": "imagem",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cargo",
                        "in": "query",
                        "description": "cargo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salario",
                        "in": "query",
                        "description": "salario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genero",
                        "in": "query",
                        "description": "genero",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cep",
                        "in": "query",
                        "description": "cep",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "logradouro",
                        "in": "query",
                        "description": "logradouro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bairro",
                        "in": "query",
                        "description": "bairro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "numero",
                        "in": "query",
                        "description": "numero",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cidade",
                        "in": "query",
                        "description": "cidade",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "estado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pais",
                        "in": "query",
                        "description": "pais",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Form request validation error. Invalid input."
                    }
                }
            },
            "post": {
                "tags": [
                    "Funcionarios"
                ],
                "description": "Store a new user into database",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "nome",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "imagem",
                        "in": "query",
                        "description": "imagem",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cargo",
                        "in": "query",
                        "description": "cargo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salario",
                        "in": "query",
                        "description": "salario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genero",
                        "in": "query",
                        "description": "genero",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cep",
                        "in": "query",
                        "description": "cep",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "logradouro",
                        "in": "query",
                        "description": "logradouro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bairro",
                        "in": "query",
                        "description": "bairro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "numero",
                        "in": "query",
                        "description": "numero",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cidade",
                        "in": "query",
                        "description": "cidade",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "estado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pais",
                        "in": "query",
                        "description": "pais",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Form request validation error. Invalid input."
                    }
                }
            }
        },
        "/api/funcionarios/{funcionario}": {
            "get": {
                "tags": [
                    "Funcionarios"
                ],
                "description": "Get a specifc user from API",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Form request validation error. Invalid input."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Funcionarios"
                ],
                "description": "Destroy a specifc user from API",
                "operationId": "destroy",
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Form request validation error. Invalid input."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Funcionarios",
            "description": "Employees API"
        }
    ]
}